@@ -8,7 +8,7 @@ version = "2024.1"
 description = "Get Descriptions of Array Computations via Lazy Evaluation"
 readme = "README.rst"
 license = "MIT"
-requires-python = "~=3.8"
+requires-python = "~=3.10"
 authors = [
     { name = "Andreas Kloeckner", email = "inform@tiker.net" },
     { name = "Matt Wala" },
