@@ -34,7 +34,8 @@ THE SOFTWARE.
 """
 
 import logging
-from typing import TYPE_CHECKING, Any, Hashable, Mapping
+from collections.abc import Hashable, Mapping
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
