@@ -29,7 +29,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from typing import Mapping
+from collections.abc import Mapping
 
 from pytato.array import (
     AbstractResultWithNamedArrays,
