@@ -3,7 +3,8 @@ from __future__ import annotations
 import operator
 import random
 import types
-from typing import Any, Callable, Sequence
+from collections.abc import Callable, Sequence
+from typing import Any
 
 import numpy as np
 
