@@ -33,8 +33,9 @@ Code Generation Targets
 .. autoclass:: BoundProgram
 """
 
+from collections.abc import Mapping
 from dataclasses import dataclass
-from typing import Any, Mapping
+from typing import Any
 
 
 class Target:
